[refactor] transactionTemplate > AOP
